+++ src/main/java/com/apap/tugas1/service/InstansiService.java
@@ -3,10 +3,9 @@ package com.apap.tugas1.service;
 import com.apap.tugas1.model.InstansiModel;
 
 import java.util.List;
-import java.math.BigInteger;
 
 public interface InstansiService {
-	InstansiModel getDetailById(BigInteger id);
+	InstansiModel getDetailById(long id);
 	List<InstansiModel> getAllInstansi();
 }
