+++ src/main/java/com/apap/tugas1/service/ProvinsiService.java
@@ -3,10 +3,9 @@ package com.apap.tugas1.service;
 import com.apap.tugas1.model.ProvinsiModel;
 
 import java.util.List;
-import java.math.BigInteger;
 
 public interface ProvinsiService {
-	ProvinsiModel getDetailById(BigInteger id);
+	ProvinsiModel getDetailById(long id);
 	List<ProvinsiModel> getAllProvinsi();
 }
